+++ src/components/payment/Confirm/components/index.js
@@ -3,5 +3,15 @@ import DeliverySection from './DeliverySection';
 import NoteSection from './NoteSection';
 import StoreInfoSection from './StoreInfoSection';
 import PricingAndPromotionSection from './PricingAndPromotionSection';
+import CommissionsSection from './CommissionsSection';
+import ActionButtonSection from './ActionButtonSection';
 
-export {PaymentMethodSection, DeliverySection, NoteSection, StoreInfoSection, PricingAndPromotionSection};
+export {
+  PaymentMethodSection,
+  DeliverySection,
+  NoteSection,
+  StoreInfoSection,
+  PricingAndPromotionSection,
+  CommissionsSection,
+  ActionButtonSection,
+};